requirements.txt
Flask==0.11.1
Flask_Cors==3.0.2
wit==4.2.0
httplib2==0.10.3
ufal.morphodita==1.9.2.1
requests==2.11.1
Werkzeug==0.11.11
beautifulsoup4==4.5.3
ordereddict==1.1
PyYAML==3.12
gunicorn==19.4.5
itsdangerous==0.24
click==6.6
Jinja2==2.8
MarkupSafe==0.23
six==1.10.0
bs4==0.0.1